# docs/api/whatsapp-ai-openapi.yaml
openapi: 3.0.3
info:
  title: autolumiku WhatsApp AI API
  description: API for WhatsApp AI Automation - customer chat & staff commands management
  version: 1.0.0
  contact:
    name: autolumiku Support
    email: support@autolumiku.com
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
servers:
  - url: https://api.autolumiku.com/v1
    description: Production server
  - url: https://staging-api.autolumiku.com/v1
    description: Staging server
  - url: http://localhost:3000/api/v1
    description: Development server
security:
  - BearerAuth: []
paths:
  /whatsapp-ai/initialize:
    post:
      tags:
        - Setup
      summary: Initialize WhatsApp connection
      description: Generate QR code for WhatsApp Business connection via Aimeow
      operationId: initializeWhatsApp
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - tenantId
              properties:
                tenantId:
                  type: string
                  format: uuid
                  example: "8dd6398e-b2d2-4724-858f-ef9cfe6cd5ed"
      responses:
        '200':
          description: QR code generated successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  data:
                    type: object
                    properties:
                      clientId:
                        type: string
                        example: "tenant_8dd6398e_1732789123456"
                      qrCode:
                        type: string
                        description: Base64 encoded QR code image
                      qrCodeExpiresAt:
                        type: string
                        format: date-time
        '400':
          $ref: '#/components/responses/BadRequest'
        '500':
          $ref: '#/components/responses/InternalServerError'
...
  /whatsapp-ai/config:
    get:
      tags:
        - Configuration
      summary: Get AI configuration
      description: Retrieve WhatsApp AI settings (personality, business hours, etc.)
      operationId: getAIConfig
      parameters:
        - name: tenantId
          in: query
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Configuration retrieved successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  data:
                    $ref: '#/components/schemas/AIConfig'
        '404':
          $ref: '#/components/responses/NotFound'

    put:
      tags:
        - Configuration
      summary: Update AI configuration
      description: Update WhatsApp AI settings
      operationId: updateAIConfig
      requestBody:
        required: true
        content:
          application/json:
            schema:
              allOf:
                - type: object
                  required:
                    - tenantId
                  properties:
                    tenantId:
                      type: string
                      format: uuid
                - $ref: '#/components/schemas/AIConfigUpdate'
      responses:
        '200':
          description: Configuration updated successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  message:
                    type: string
                  data:
                    $ref: '#/components/schemas/AIConfig'
        '400':
          $ref: '#/components/responses/BadRequest'
        '404':
          $ref: '#/components/responses/NotFound'

    post:
      tags:
        - Configuration
      summary: Create initial AI configuration
      description: Create AI configuration during setup (called automatically)
      operationId: createAIConfig
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - tenantId
                - accountId
              properties:
                tenantId:
                  type: string
                  format: uuid
                accountId:
                  type: string
                  format: uuid
                aiName:
                  type: string
                  default: "AI Assistant"
                welcomeMessage:
                  type: string
      responses:
        '200':
          description: Configuration created successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  data:
                    $ref: '#/components/schemas/AIConfig'
...
  /whatsapp-ai/staff:
    get:
      tags:
        - Staff Management
      summary: List staff with WhatsApp access
      description: Retrieve all staff members authorized for WhatsApp commands
      operationId: listStaff
      parameters:
        - name: tenantId
          in: query
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Staff list retrieved successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/StaffMember'

    post:
      tags:
        - Staff Management
      summary: Add staff member
      description: Grant WhatsApp command access to a staff member
      operationId: addStaff
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateStaffRequest'
      responses:
        '201':
          description: Staff member added successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  data:
                    $ref: '#/components/schemas/StaffMember'
  /whatsapp-ai/staff/{staffId}:
    put:
      tags:
        - Staff Management
      summary: Update staff permissions
      description: Modify staff member permissions for WhatsApp commands
      operationId: updateStaff
      parameters:
        - name: staffId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateStaffRequest'
      responses:
        '200':
          description: Staff updated successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  data:
                    $ref: '#/components/schemas/StaffMember'

    delete:
      tags:
        - Staff Management
      summary: Remove staff access
      description: Revoke WhatsApp command access from staff member
      operationId: deleteStaff
      parameters:
        - name: staffId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Staff removed successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  message:
                    type: string
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    AIConfig:
      type: object
      properties:
        id:
          type: string
          format: uuid
        tenantId:
          type: string
          format: uuid
        accountId:
          type: string
          format: uuid
        aiName:
          type: string
          example: "AutoBot"
        aiPersonality:
          type: string
          enum: [friendly, professional, casual]
          default: friendly
        welcomeMessage:
          type: string
          example: "Halo! 👋 Ada yang bisa saya bantu?"
        autoReply:
          type: boolean
          default: true
        customerChatEnabled:
          type: boolean
          default: true
        staffCommandsEnabled:
          type: boolean
          default: true
        businessHours:
          type: object
          example:
            monday: { open: "09:00", close: "17:00" }
            tuesday: { open: "09:00", close: "17:00" }
        timezone:
          type: string
          example: "Asia/Jakarta"
        afterHoursMessage:
          type: string
        aiModel:
          type: string
          example: "glm-4-plus"
        temperature:
          type: number
          format: float
          minimum: 0
          maximum: 1
          default: 0.7
        maxTokens:
          type: integer
          default: 1000
        customFAQ:
          type: object
        productKnowledge:
          type: object
        enableVehicleInfo:
          type: boolean
          default: true
        enablePriceNegotiation:
          type: boolean
          default: false
        enableTestDriveBooking:
          type: boolean
          default: true
        enableStaffUpload:
          type: boolean
          default: true
        enableStaffStatus:
          type: boolean
          default: true
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
    AIConfigUpdate:
      type: object
      properties:
        aiName:
          type: string
        aiPersonality:
          type: string
          enum: [friendly, professional, casual]
        welcomeMessage:
          type: string
        autoReply:
          type: boolean
        customerChatEnabled:
          type: boolean
        staffCommandsEnabled:
          type: boolean
        businessHours:
          type: object
        timezone:
          type: string
        afterHoursMessage:
          type: string
        temperature:
          type: number
        maxTokens:
          type: integer
...
    StaffMember:
      type: object
      properties:
        id:
          type: string
          format: uuid
        tenantId:
          type: string
          format: uuid
        phoneNumber:
          type: string
          example: "+628123456789"
        staffName:
          type: string
          example: "John Doe"
        canUploadVehicle:
          type: boolean
          default: false
        canUpdateStatus:
          type: boolean
          default: false
        canViewInventory:
          type: boolean
          default: true
        canViewStats:
          type: boolean
          default: true
        isActive:
          type: boolean
          default: true
        createdAt:
          type: string
          format: date-time
    CreateStaffRequest:
      type: object
      required:
        - tenantId
        - phoneNumber
        - staffName
      properties:
        tenantId:
          type: string
          format: uuid
        phoneNumber:
          type: string
          pattern: "^\\+628[0-9]{8,12}$"
        staffName:
          type: string
        canUploadVehicle:
          type: boolean
          default: false
        canUpdateStatus:
          type: boolean
          default: false
    UpdateStaffRequest:
      type: object
      properties:
        staffName:
          type: string
        canUploadVehicle:
          type: boolean
        canUpdateStatus:
          type: boolean
        canViewInventory:
          type: boolean
        canViewStats:
          type: boolean
        isActive:
          type: boolean
...
  responses:
    BadRequest:
      description: Bad request - validation failed
      content:
        application/json:
          schema:
            type: object
            properties:
              success:
                type: boolean
                example: false
              error:
                type: string
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            type: object
            properties:
              success:
                type: boolean
                example: false
              error:
                type: string
    InternalServerError:
      description: Internal server error
      content:
        application/json:
          schema:
            type: object
            properties:
              success:
                type: boolean
                example: false
              error:
                type: string
              details:
                type: string
...